create invalidUser exceptio

+++ src/main/java/com/usersDb/usersDb/service/UserService.java
@@ -1,6 +1,8 @@
 package com.usersDb.usersDb.service;
 
+import com.usersDb.usersDb.exception.AlreadyExistEmailException;
 import com.usersDb.usersDb.exception.FindByNameException;
+import com.usersDb.usersDb.exception.InvalidUserException;
 import com.usersDb.usersDb.exception.UserCreationException;
 import com.usersDb.usersDb.model.User;
 import com.usersDb.usersDb.repository.UserRepository;
@@ -8,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.io.FileNotFoundException;
 import java.util.List;
 
 @Service
@@ -17,9 +18,9 @@ public class UserService {
     UserRepository userRepository;
 
     public User createUser(User user) {
-        if (user.getName().equals("") || user.getPassword().equals("")) throw new UserCreationException();
+        if (user.getName().equals("") || user.getEmail().equals("") || user.getPassword().equals("")) throw new UserCreationException();
+        if(this.userRepository.existsByEmail(user.getEmail())) throw new AlreadyExistEmailException("Already exist this email");
         return this.userRepository.save(user);
-
 
     }
 
@@ -28,6 +29,7 @@ public class UserService {
     }
 
     public void deleteUserById(final Long userId) {
+        if (!userRepository.existsById(userId)) throw new InvalidUserException();
         this.userRepository.deleteById(userId);
     }
 
@@ -44,11 +46,13 @@ public class UserService {
         return userRepository.save(updatedUser);
     }
 
+
     public User updateUserDetail(Long userId, User user) {
         User updatedUser = userRepository.getReferenceById(userId);
-        if(user.getName() != null) {
+        if (user.getName() != null) {
             updatedUser.setName(user.getName());
-        } if (user.getPassword() !=null) {
+        }
+        if (user.getPassword() != null) {
             updatedUser.setPassword(user.getPassword());
         }
         return userRepository.save(updatedUser);
